backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
celery==5.4.0
django-celery-beat==2.6.0
redis==5.0.4
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==24.2
requests==2.32.3
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.6.0
eventlet==0.36.1
pytest==8.2.0
pytest-django==4.8.0
pytest-asyncio==0.23.7
